+++ RELEASENOTES.md
@@ -18,6 +18,8 @@
 *   Extractors:
 *   DataSource:
 *   Audio:
+    *   Do not bypass `SonicAudioProcessor` when `SpeedChangingAudioProcessor`
+        is configured with default parameters.
 *   Video:
 *   Text:
 *   Metadata: